Move new media implementation into media file and delete unused camera file
